# backend/openapi.yaml
openapi: 3.0.3
info:
  title: TradeIQ API
  version: 0.1.0
  description: >
    Phase 1 (complete). Implemented: /auth, /outlets, /skus, /visits (check-in
    + submit + list), /stock (S2), /visibility (S3-S4, vision-stub wired),
    /pricing (S5, OCR-stub wired), /competitive (S6), /capability (S7), /risks
    (S8), /tasks (S9), /scorecards (S10), /dashboard, and /photos (photo
    capture + photo-verified task closure). Every mounted route is implemented
    — no route returns 501. Manager-facing GET listings on the per-visit
    section modules follow a common contract: GET /<section>?visitId=<id>
    returns that visit's captured rows (any authenticated role, tenant-scoped,
    400 if visitId is missing, 404 if the visit is not the caller's). All
    non-auth/non-health routes require a Bearer JWT; write routes are
    role-guarded (403 when the caller's role is not permitted).
    Phase 3 (Activation) adds /territories, /campaigns, /beatplans, /alerts,
    /templates, and /trends — see docs/architecture/phase3-activation.md.
    Phase 2 (Intelligence) adds /fraud, /dispatch, /forecast; Phase 3 Tier-2
    adds /orders, /messages, /announcements, /webhooks, /gamification, /reports
    — see docs/architecture/phase2-intelligence.md.
servers:
  - url: http://localhost:4000
paths:
...
  /territories:
    get:
      summary: List the client's territories
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Territories }, '401': { description: Unauthorized } }
    post:
      summary: Create a territory — manager/admin
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Created }, '400': { description: Invalid }, '403': { description: Forbidden }, '409': { description: Duplicate code } }
  /territories/{id}/agents:
    post:
      summary: Assign an agent to a territory — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '201': { description: Assigned }, '404': { description: Territory or user not found }, '409': { description: Already assigned } }
  /territories/{id}/coverage:
    get:
      summary: Outlets + assigned agents for a territory
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Coverage }, '404': { description: Not found } }
...
  /fraud/visits/{visitId}:
    get:
      summary: Fraud risk score + signals for a visit — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: visitId, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 'Risk score + signals' }, '404': { description: Not found } }
  /fraud/attempts:
    get:
      summary: List check-in attempts (incl. rejected) — manager/admin
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: outletId, in: query, schema: { type: string } }
        - { name: agentId, in: query, schema: { type: string } }
        - { name: passed, in: query, schema: { type: string, enum: ['true', 'false'] } }
      responses: { '200': { description: Attempts } }
  /fraud/flagged:
    get:
      summary: Submitted visits above a fraud-risk threshold — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: minScore, in: query, schema: { type: number } }]
      responses: { '200': { description: Flagged visits } }
  /dispatch:
    post:
      summary: Rank agents for an outlet (nearest + in-territory) — manager/admin
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: 'Ranked candidates + recommended' }, '400': { description: Missing outletId }, '404': { description: Outlet not found } }
  /dispatch/agents:
    get:
      summary: Ranked agent candidates for an outlet — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: outletId, in: query, required: true, schema: { type: string } }]
      responses: { '200': { description: Candidates }, '400': { description: Missing outletId } }
...
  /webhooks:
    get:
      summary: List webhook subscriptions — manager/admin
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Webhooks } }
    post:
      summary: Create a webhook subscription — manager/admin
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Created }, '400': { description: Invalid url/event }, '403': { description: Forbidden } }
  /webhooks/{id}:
    patch:
      summary: Update a webhook — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Updated }, '400': { description: Invalid }, '404': { description: Not found } }
    delete:
      summary: Delete a webhook — manager/admin
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '204': { description: Deleted }, '404': { description: Not found } }
  /gamification/leaderboard:
    get:
      summary: Field-agent leaderboard (points from scorecards/tasks/visits)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses: { '200': { description: Leaderboard }, '400': { description: Invalid dates } }
  /gamification/me:
    get:
      summary: The caller's own leaderboard entry + rank
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Entry } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT